Add method for getting Room and user access status.
RoomManager.get_access_status will return a tuple with two values. The
first will be the instance of Room (or None if not found). The second
will be an instance of RoomAccessStatus, an Enum that provides
constants indicating whether or not the user can join a room and why.
RoomAccessStatus might have to be updated in the future if there are
additional reasons why a user isn't allowed to join a room.